requirements.txt
gunicorn
flask
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
tox
psycopg2-binary
raven
blinker
redis
rq
boto
flask-restx
enum34
barcodenumber
requests
newrelic
flask_cors
factory-boy
pytest-factoryboy
APScheduler
backoff
dateparser
sendgrid
